+++ pyproject.toml
@@ -203,6 +203,7 @@ select = [
     "F",      # pyflakes
     "I",      # isort
     "PERF",   # Perflint
+    "RUF022", # unsorted-dunder-all: keep `__all__` sorted
     "UP",     # pyupgrade
     "TID251", # https://docs.astral.sh/ruff/rules/banned-api/
 ]
@@ -219,6 +220,8 @@ max-complexity = 24 # Default is 10
 "__init__.py" = ["F401"]
 # The mcp.types package is an alias that mirrors mcp_types namespaces by design.
 "src/mcp/types/*.py" = ["F403"]
+# mcp_types' __all__ is grouped by section comments (capabilities, content types, requests, ...) rather than sorted.
+"src/mcp-types/mcp_types/__init__.py" = ["RUF022"]
 # Generated by scripts/gen_surface_types.py: raw datamodel-codegen output (TID251 lifts the repo-wide RootModel ban for these generated validators).
 "src/mcp-types/mcp_types/_v*/__init__.py" = ["D212", "E501", "I001", "TID251", "UP007", "UP037"]
 "tests/server/mcpserver/test_func_metadata.py" = ["E501"]
