+++ src/mcp-types/mcp_types/methods.py
@@ -27,7 +27,6 @@ __all__ = [
     "CLIENT_NOTIFICATIONS",
     "CLIENT_REQUESTS",
     "CLIENT_RESULTS",
-    "CacheableMethod",
     "INPUT_REQUIRED_METHODS",
     "MONOLITH_NOTIFICATIONS",
     "MONOLITH_REQUESTS",
@@ -37,6 +36,7 @@ __all__ = [
     "SERVER_RESULTS",
     "SPEC_CLIENT_METHODS",
     "SPEC_CLIENT_NOTIFICATION_METHODS",
+    "CacheableMethod",
     "is_input_required",
     "parse_client_notification",
     "parse_client_request",
