+++ src/mcp-types/mcp_types/version.py
@@ -10,14 +10,14 @@ ordering questions go through KNOWN_PROTOCOL_VERSIONS.
 from typing import Final
 
 __all__ = [
-    "KNOWN_PROTOCOL_VERSIONS",
     "HANDSHAKE_PROTOCOL_VERSIONS",
-    "MODERN_PROTOCOL_VERSIONS",
-    "SUPPORTED_PROTOCOL_VERSIONS",
-    "LATEST_PROTOCOL_VERSION",
+    "KNOWN_PROTOCOL_VERSIONS",
     "LATEST_HANDSHAKE_VERSION",
     "LATEST_MODERN_VERSION",
+    "LATEST_PROTOCOL_VERSION",
+    "MODERN_PROTOCOL_VERSIONS",
     "OLDEST_SUPPORTED_VERSION",
+    "SUPPORTED_PROTOCOL_VERSIONS",
     "is_version_at_least",
 ]
