+++ src/mcp/cli/__init__.py
@@ -2,5 +2,4 @@
 
 from .cli import app
 
-if __name__ == "__main__":  # pragma: no cover
-    app()
+__all__ = ["app"]
