+++ src/mcp/server/__init__.py
@@ -4,4 +4,4 @@ from .lowlevel import NotificationOptions, Server
 from .mcpserver import MCPServer
 from .models import InitializationOptions
 
-__all__ = ["CacheHint", "Server", "ServerRequestContext", "MCPServer", "NotificationOptions", "InitializationOptions"]
+__all__ = ["CacheHint", "InitializationOptions", "MCPServer", "NotificationOptions", "Server", "ServerRequestContext"]
