+++ src/mcp/server/auth/handlers/register.py
@@ -14,10 +14,6 @@ from mcp.server.auth.provider import OAuthAuthorizationServerProvider, Registrat
 from mcp.server.auth.settings import ClientRegistrationOptions
 from mcp.shared.auth import JWT_BEARER_GRANT_TYPE, OAuthClientInformationFull, OAuthClientMetadata
 
-# this alias is a no-op; it's just to separate out the types exposed to the
-# provider from what we use in the HTTP handler
-RegistrationRequest = OAuthClientMetadata
-
 
 class RegistrationErrorResponse(BaseModel):
     error: RegistrationErrorCode
