+++ src/mcp/server/mcpserver/__init__.py
@@ -28,30 +28,30 @@ from .server import MCPServer, require_client_extension
 from .utilities.types import Audio, Image
 
 __all__ = [
-    "MCPServer",
-    "Context",
-    "Image",
-    "Audio",
-    "Icon",
-    "Resolve",
-    "Elicit",
-    "Sample",
-    "ListRoots",
-    "ElicitationResult",
-    "AcceptedElicitation",
-    "DeclinedElicitation",
-    "CancelledElicitation",
-    "Extension",
-    "ToolBinding",
-    "ResourceBinding",
-    "MethodBinding",
-    "require_client_extension",
-    "ResourceSecurity",
     "DEFAULT_RESOURCE_SECURITY",
-    "RequestStateSecurity",
-    "RequestStateCodec",
-    "RequestStateBoundary",
     "AESGCMRequestStateCodec",
+    "AcceptedElicitation",
+    "Audio",
+    "CancelledElicitation",
+    "Context",
+    "DeclinedElicitation",
+    "Elicit",
+    "ElicitationResult",
+    "Extension",
+    "Icon",
+    "Image",
     "InvalidRequestState",
+    "ListRoots",
+    "MCPServer",
+    "MethodBinding",
+    "RequestStateBoundary",
+    "RequestStateCodec",
+    "RequestStateSecurity",
+    "Resolve",
+    "ResourceBinding",
+    "ResourceSecurity",
+    "Sample",
+    "ToolBinding",
     "authenticated_principal",
+    "require_client_extension",
 ]
