+++ src/mcp/shared/inbound.py
@@ -38,14 +38,14 @@ from mcp_types.version import MODERN_PROTOCOL_VERSIONS
 
 __all__ = [
     "ERROR_CODE_HTTP_STATUS",
-    "InboundLadderRejection",
-    "InboundModernRoute",
     "MCP_METHOD_HEADER",
     "MCP_NAME_HEADER",
     "MCP_PARAM_HEADER_PREFIX",
     "MCP_PROTOCOL_VERSION_HEADER",
     "NAME_BEARING_METHODS",
     "X_MCP_HEADER_KEY",
+    "InboundLadderRejection",
+    "InboundModernRoute",
     "classify_inbound_request",
     "decode_header_value",
     "encode_header_value",
